feat: NoteCard
